requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.20
django-embed-video==1.3.1
idna==3.4
Pillow==8.4.0
psycopg2==2.9.6
pytz==2023.3
requests==2.27.1
sqlparse==0.4.4
typing-extensions==4.1.1
urllib3==1.26.16
